bug fixed

+++ Task06/Program.cs
@@ -3,6 +3,6 @@
 // -3 -> нет
 // 7 -> нет
 Console.WriteLine("Введите число: ");
-int numberA = Convert.ToInt16(Console.ReadLine());
+int numberA = Convert.ToInt32(Console.ReadLine());
 if (numberA % 2 == 0) Console.WriteLine($"Число {numberA} является четным");
 else Console.WriteLine($"Число {numberA} является нечетным");
